+++ src/Parser/Paladin/Retribution/CONFIG.js
