+++ src/Parser/Paladin/Retribution/Modules/HolyPower/HolyPowerDetails.js
