+++ src/Parser/Paladin/Retribution/Modules/Items/Tier20_2set.js
@@ -8,7 +8,7 @@ import { formatNumber, formatPercentage } from 'common/format';
 import Module from 'Parser/Core/Module';
 import Combatants from 'Parser/Core/Modules/Combatants';
 
-import GetDamageBonus from '../PaladinCore/GetDamageBonus';
+import GetDamageBonus from '../../../Shared/GetDamageBonus';
 
 const RET_PALADIN_T20_2SET_MODIFIER = 0.2;
