+++ src/Parser/Paladin/Retribution/Modules/PaladinCore/Haste.js
