+++ src/Parser/Paladin/Retribution/Modules/Traits/DeliverTheJustice.js
@@ -7,7 +7,7 @@ import SPELLS from 'common/SPELLS';
 import SpellIcon from 'common/SpellIcon';
 import SpellLink from 'common/SpellLink';
 import { formatPercentage } from 'common/format';
-import getDamageBonusStacked from '../PaladinCore/getDamageBonusStacked';
+import getDamageBonusStacked from '../../../Shared/getDamageBonusStacked';
 
 const DELIVER_THE_JUSTICE_INCREASE = 0.08;
 /**
