+++ src/Parser/Paladin/Retribution/Modules/Traits/RelicTraits.js
