+++ src/Parser/Paladin/Retribution/Modules/Traits/WrathOfTheAshbringer.js
