+++ src/Parser/Paladin/Shared/GetDamageBonus.js
