+++ src/Parser/Paladin/Shared/getDamageBonusStacked.js
